Refactor method to create or update state data
